modified:   hadoop-yarn-project/hadoop-yarn/hadoop-yarn-common/src/test/java/org/apache/hadoop/yarn/util/resource/TestResources.java

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-common/src/test/java/org/apache/hadoop/yarn/util/resource/TestResources.java
@@ -24,7 +24,7 @@ public class TestResources {
         
         Assert.assertTrue(Resources.fitsIn(rhs, lhs));
              
-        int allcatedGPU = Resources.allocateGPUs(rhs, lhs);
+        long allcatedGPU = Resources.allocateGPUs(rhs, lhs);
         Assert.assertEquals(allcatedGPU, 3);
         
         ret = Resources.add(ret, rhs);
